add avatar and cover image to cloudinary and bug fixing

+++ youtube-clone-fullstack/src/utils/Cloudinary.js
@@ -4,8 +4,8 @@ import fs from "fs";
 //upload image to cloudinary
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret: process.env.CLOUDINARY_API_SECRET,
+    api_key: process.env.CLOUDINARY_CLOUD_API_KEY,
+    api_secret: process.env.CLOUDINARY_CLOUD_API_SECRET,
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
@@ -17,17 +17,26 @@ const uploadOnCloudinary = async (localFilePath) => {
             resource_type: "auto",
         });
 
+        //testing purpose only
+        // console.log("response", response);
+
         //file has been uploaded to cloudinary
-        console.log(
-            "File uploaded to cloudinary, deleting local file...",
-            response.url,
-        );
+        //for testing purpose we are logging the url of the uploaded file
+        // console.log(
+        //     "File uploaded to cloudinary, deleting local file...",
+        //     response.url,
+        // );
+
+        if (localFilePath) {
+            //remove the locally saved temporary file
+            fs.unlinkSync(localFilePath);
+        }
 
         return response;
     } catch (error) {
-        // remove the locally saved temporary file as the upload operation got failed
-        fs.unlinkSync(localFilePath);
+            // remove the locally saved temporary file as the upload operation got failed
+            fs.unlinkSync(localFilePath);
     }
 };
 
-export default uploadOnCloudinary;
+export { uploadOnCloudinary };
